requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Time series and forecasting
prophet==1.1.4
statsmodels==0.14.0
tsfresh==0.20.1

# Deep learning
tensorflow==2.13.0
torch==2.0.1
keras==2.13.1

# Data processing
pyspark==3.4.1
dask==2023.7.1
apache-kafka==2.8.2

# AWS services
boto3==1.28.25
sagemaker==2.185.0
kinesis-python==0.1.0

# MLOps
mlflow==2.5.0
dvc[s3]==3.0.0
wandb==0.15.8

# Data visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
bokeh==3.2.2

# Monitoring and alerting
prometheus-client==0.17.1
grafana-api==1.0.3

# Web framework
fastapi==0.101.1
uvicorn==0.23.2
streamlit==1.25.0

# Database
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
redis==4.6.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Development tools
jupyter==1.0.0
jupyterlab==4.0.7
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.65.0
joblib==1.3.2

# Infrastructure
terraform-compliance==1.3.44
cdktf==0.17.0
